Bug fix. Theater found optimally

+++ src/main/java/com/chobichokro/controllers/FeatureController.java
@@ -11,11 +11,13 @@ import com.chobichokro.repository.TheaterRepository;
 import com.chobichokro.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/feature")
@@ -30,47 +32,64 @@ public class FeatureController {
     TheaterRepository theaterRepository;
     @Autowired
     TheaterHelper theaterHelper;
+
     @PutMapping("/make-all-id-same")
+//    @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<?> makeAllOwnerIDToLicenseId() {
-        // schedule done
-       List<Schedule> scheduleList = scheduleRepository.findAll();
-       for(Schedule schedule : scheduleList){
-           String theaterID = schedule.getTheaterId();
-           schedule.setTheaterId(getTheaterOwner(theaterID));
-           scheduleRepository.save(schedule);
-       }
+//         schedule done
+        System.out.println("schedule");
+        List<Schedule> scheduleList = scheduleRepository.findAll();
+        for (Schedule schedule : scheduleList) {
+            String theaterID = schedule.getTheaterId();
+            schedule.setTheaterId(getTheaterOwner(theaterID));
+            scheduleRepository.save(schedule);
+        }
 
-       // theater done
-       List<Theater> theatersList = theaterRepository.findAll();
-       for(Theater theater : theatersList){
-           String theaterOwner = getTheaterOwner(theater.getId());
-           theaterRepository.delete(theater);
-           theater.setLicenseId(theaterOwner);
-           theater.setId(theaterOwner);
-           theaterRepository.save(theater);
+//         theater done
+        System.out.println("theater");
 
-       }
+        List<Theater> theatersList = theaterRepository.findAll();
+        for (Theater theater : theatersList) {
+            System.out.println(theater);
+            String theaterOwner = getTheaterOwner(theater.getLicenseId());
+            if(theaterOwner == null) continue;
+            theaterRepository.delete(theater);
+            theater.setLicenseId(theaterOwner);
+            theater.setId(theaterOwner);
+            theaterRepository.save(theater);
 
-       // license done
+        }
 
-       List<License> licensesList = licenseRepository.findAll();
-       licenseRepository.deleteAll(licensesList);
-       for(License license : licensesList){
-           String licenseOwner = license.getLicenseOwner();
-           license.setId(licenseOwner);
-           licenseRepository.save(license);
-       }
+        // license done
+        System.out.println("license");
 
-       List<User> userList = userRepository.findAll();
-       for(User user : userList){
-           user.setLicenseId(user.getId());
-           userRepository.save(user);
-       }
-       return ResponseEntity.ok(userList.toString() + licensesList.toString() + theatersList.toString() + scheduleList.toString());
+        List<License> licensesList = licenseRepository.findAll();
+        licenseRepository.deleteAll(licensesList);
+        for (License license : licensesList) {
+            String licenseOwner = license.getLicenseOwner();
+            license.setId(licenseOwner);
+            licenseRepository.save(license);
+        }
+        System.out.println("user");
+
+        List<User> userList = userRepository.findAll();
+        for (User user : userList) {
+            user.setLicenseId(user.getId());
+            userRepository.save(user);
+        }
+        return ResponseEntity.ok(userList.toString() + licensesList.toString() + theatersList.toString());
     }
-    public String getTheaterOwner(String theaterId){
-        String licenceID = theaterRepository.findById(theaterId).get().getLicenseId();
-        return licenseRepository.findById(licenceID).get().getLicenseOwner();
+
+    public String getTheaterOwner(String licenseId) {
+        Optional<License> license = licenseRepository.findLicenseById(licenseId);
+        if(license.isPresent()){
+            System.out.println(license);
+        }
+        else{
+            System.out.println("no license");
+            return null;
+        }
+        return licenseRepository.findById(licenseId).get().getLicenseOwner();
     }
 
 }
